[FEAT] 인증 회원가입 추가

+++ src/main/java/groom/him/core/auth/util/interceptor/RefreshValidateInterceptor.java
@@ -32,7 +32,7 @@ public class RefreshValidateInterceptor implements HandlerInterceptor {
         String refreshToken = jwtTokenProvider.resolveToken(request);
         try {
             if (refreshToken != null) {
-                Long userId = Long.parseLong(jwtTokenProvider.getUserId(refreshToken));
+                Integer userId = Integer.parseInt(jwtTokenProvider.getUserId(refreshToken));
                 if (authService.existsRefreshToken(userId, refreshToken) && jwtTokenProvider.isTokenNonExpired(refreshToken)) {
                     try {
                         Authentication authentication = getAuthentication(refreshToken);
